+++ benlibc/syscall.h
@@ -4,6 +4,5 @@ void syscall_printchar(int c);
 void syscall_printstr(char* s);
 void syscall_cls(void);
 int syscall_fread(char* path, char* buf, int size);
-void syscall_lsdir(char* path, char* buf);
+int syscall_lsdir(char* path, char* buf);
 int syscall_getchar(void);
-int syscall_ffind(char* path);